fix norma: exec/redirect
