--- f3
@@ -1,18 +1,0 @@
-f1
-f2
-f3
-includes
-libft
-log
-Makefile
-minishell
-minishell.excalidraw
-minishell_tester-main
-minishell_tester-main.zip
-mini_tests
-out
-out.txt
-readline.supp
-README.md
-src
-tests
